refactoring to prepare for MiKo_2202 and MiKo_2203 codefix providers
